backend/requirements.txt
argon2-cffi
argon2-cffi-bindings
asgiref
boto3
botocore
certifi
cffi
charset-normalizer
cryptography
defusedxml
Django
django-ckeditor
django-cors-headers
django-environ
django-js-asset
django-storages
django-templated-mail
djangorestframework
djangorestframework-simplejwt
djoser
gunicorn
idna
jmespath
oauthlib
packaging
Pillow
psycopg
psycopg2
pycparser
PyJWT
python-dateutil
python3-openid
pytz
requests
requests-oauthlib
s3transfer
six
social-auth-app-django
social-auth-core
sqlparse
typing_extensions
tzdata
urllib3
whitenoise
django-filter